+++ src/consumer/component/consumer-component.js
@@ -101,6 +101,7 @@ export type ComponentProps = {
   testerPackageDependencies?: ?Object,
   customResolvedPaths?: ?(customResolvedPath[]),
   overrides: ComponentOverrides,
+  packageJsonFile?: ?PackageJsonFile,
   files: SourceFile[],
   docs?: ?(Doclet[]),
   dists?: Dist[],
@@ -159,7 +160,7 @@ export default class Component {
   customResolvedPaths: customResolvedPath[]; // used when in the same component, one file requires another file using custom-resolve
   _driver: Driver;
   _isModified: boolean;
-  packageJsonFile: PackageJsonFile;
+  packageJsonFile: ?PackageJsonFile; // populated when loadedFromFileSystem or when writing the components. for author it never exists
   _currentlyUsedVersion: BitId; // used by listScope functionality
   pendingVersion: Version; // used during tagging process. It's the version that going to be saved or saved already in the model
   dataToPersist: DataToPersist;
@@ -213,6 +214,7 @@ export default class Component {
     compilerPackageDependencies,
     testerPackageDependencies,
     overrides,
+    packageJsonFile,
     docs,
     dists,
     mainDistFile,
@@ -247,6 +249,7 @@ export default class Component {
     this.compilerPackageDependencies = compilerPackageDependencies || {};
     this.testerPackageDependencies = testerPackageDependencies || {};
     this.overrides = overrides;
+    this.packageJsonFile = packageJsonFile;
     this._docs = docs;
     this.setDists(dists, mainDistFile ? path.normalize(mainDistFile) : null);
     this.specsResults = specsResults;
@@ -1108,6 +1111,7 @@ export default class Component {
       componentConfig,
       isAuthor
     );
+    const packageJsonFile = (componentConfig && componentConfig.packageJsonFile) || null;
 
     return new Component({
       name: id.name,
@@ -1128,7 +1132,8 @@ export default class Component {
       testerPackageDependencies,
       deprecated,
       origin: componentMap.origin,
-      overrides
+      overrides,
+      packageJsonFile
     });
   }
 }